+++ source/js/modules/timer.js
