+++ source/js/script.js
@@ -12,7 +12,7 @@ import loadPage from './modules/page-load';
 import rules from './modules/rules.js';
 import LetterAnimation from './modules/letter-animation.js';
 import prizes from './modules/prizes.js';
-import counter from './modules/counter.js';
+import timer from './modules/timer.js';
 
 // init modules
 mobileHeight();
@@ -26,7 +26,7 @@ social();
 loadPage();
 rules();
 prizes();
-counter();
+timer();
 
 const fullPageScroll = new FullPageScroll();
 fullPageScroll.init();
